netapp-cli/pyproject.toml
[project]
name = "netapp-cli"
version = "0.1.0"
description = "CLI tool to interact with ActiveIQ Server for NetApp storage management and automation"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "NetApp CLI", email = "admin@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Systems Administration",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.1.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "tabulate>=0.9.0",
    "python-dateutil>=2.8.0",
]

[project.scripts]
netapp = "netapp_cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
select = ["E", "F", "I", "N", "W", "UP"]
line-length = 88
target-version = "py311"
